requirements.txt
azure-core==1.30.0
azure-iot-hub==2.6.1
click==8.1.7
docutils==0.20.1
Flask==3.0.2
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
msrest==0.7.1
oauthlib==3.2.2
Pygments==2.17.2
requests==2.31.0
requests-oauthlib==1.3.1
uamqp==1.6.8
urllib3==2.2.0
